feat(cangjie_runtime): sync from upstream v1.1.0-alpha.70

+++ runtime/src/TypeInfoManager.cpp
@@ -612,7 +612,7 @@ void TypeInfoManager::CalculateGCTib(TypeInfo* typeInfo)
     CString gcTibStr = typeGCInfo.GetGCTibStr(typeInfo);
     size_t len = gcTibStr.Length();
     GCTib gcTib;
-    constexpr uint8_t alignSize = sizeof(uint32_t);
+    constexpr uint8_t alignSize = sizeof(uint64_t);
     // create StdGCTib
     U16 num = gcTibStr.Length() / alignSize;
     U16 needSpace = sizeof(U32) + sizeof(U8) * num;
